Modified introduction and setup pages structure (again)
Intro page now really explains the fastest way to get up and running with Appium on TestObject (which is through Basic setup, not suites). This is due to a change in policy: in the beginning we wanted to push all users towards the new Suite setup/approach, now it is clear we need both the Watcher setup (formerly known as Intermediate) and Suites to make all the customers satisfied.

Setup pages' structure was changed to allow more info about the setup (not strictly code) to be placed close to the setup code itself while keeping everything nice and tidy.